code modified
A new script has been added for Est6. The script could be found in the name of Script_Est6.py Please note the script is written in python instead of jupyter notebook because I was having some difficulties in running comands specific to python.pathlib

the output could be found in ..\Factors and Ridership Data\Script Outputs\FIG x-x

Please ignore the noming cleature of the figures for the time being. It were behaving very weirdly when i tried to change it.

Thank you
Vedant